+++ tests/testthat/test-docx-footnotes.R
@@ -1,29 +1,38 @@
 test_that("add footnotes", {
 
-  img.file <- file.path( R.home("doc"), "html", "logo.jpg" )
-  bold_fp <- shortcuts$fp_bold()
-  normal_fp <- fp_text()
-  bl1 <- block_list(
-    fpar(ftext("hello", normal_fp), ftext(" world!", bold_fp)),
+  fp_bold <- fp_text_lite(bold = TRUE)
+  fp_refnote <- fp_text_lite(vertical.align = "superscript")
+
+  img.file <- file.path(R.home("doc"), "html", "logo.jpg")
+  bl <- block_list(
+    fpar(ftext("hello", fp_bold)),
     fpar(
-      ftext("Do you enjoy with ", bold_fp),
-      external_img(src = img.file, height = 1.06/3, width = 1.39/3)
+      ftext("hello world", fp_bold),
+      external_img(src = img.file, height = 1.06, width = 1.39)
     )
   )
-  bl2 <- block_list(
-    fpar(ftext("hello", normal_fp), ftext(" world!", bold_fp))
+
+  a_par <- fpar(
+    "this paragraph contains a note ",
+    run_footnote(x = bl, prop = fp_refnote),
+    "."
+  )
+  b_par <- fpar(
+    "this paragraph contains another note ",
+    run_footnote(
+      x = block_list(
+        fpar(ftext("Salut", fp_text_lite()))
+      ),
+      prop = fp_refnote),
+    "."
   )
 
-  x <- read_docx()
-  x <- body_add_par(x, "Here is a footnote", style = "Normal")
-  x <- slip_in_footnote(x, style = "reference_id", blocks = bl1)
-  x <- body_add_par(x, "Here is another footnote", style = "Normal")
-  x <- slip_in_footnote(x, style = "reference_id", blocks = bl2)
+  doc <- read_docx()
+  doc <- body_add_fpar(doc, value = a_par, style = "Normal")
+  doc <- body_add_fpar(doc, value = b_par, style = "Normal")
 
-  docx_file <- tempfile(fileext = ".docx")
+  docx_file <- print(doc, target = tempfile(fileext = ".docx"))
   docx_dir <- tempfile()
-  print(x, target = docx_file)
-
   unpack_folder(docx_file, docx_dir)
 
   doc <- read_xml(file.path(docx_dir, "word/footnotes.xml"))
@@ -38,11 +47,3 @@ test_that("add footnotes", {
 
 })
 
-test_that("check blocks", {
-
-  x <- read_docx()
-  x <- body_add_par(x, "Here is a footnote", style = "Normal")
-  expect_error(slip_in_footnote(x, style = "reference_id", blocks = "hi"),
-               regexp = "block_list")
-
-})